chore: export types from sidesheet
