+++ apps/test-app/src/sidesheet/test-app-sidesheet/HandoverSidesheet.tsx
@@ -1,8 +1,9 @@
 import { Button } from '@equinor/eds-core-react';
 import { createWidget, useResizeContext } from '@equinor/workspace-sidesheet';
+import { sidesheets } from './sidesheets';
 import { render as WorkOrderRender } from './WorkorderSidesheet';
 
-type HandoverProps = {
+export type HandoverProps = {
 	id: string;
 	initial?: { id: string; description: string };
 };
@@ -17,13 +18,7 @@ export const { Component: HandoverSidesheet, render } = createWidget<HandoverPro
 	return (
 		<div>
 			<div>Am handover sidesheet</div>
-			<Button
-				onClick={() =>
-					frame.replace((el, frame) => WorkOrderRender({ el, frame, props: { woId: 'Some new id' } }))
-				}
-			>
-				Replace with WO
-			</Button>
+			<Button onClick={() => frame.replace(sidesheets.workorder({ woId: '122223' }))}>Replace with WO</Button>
 		</div>
 	);
 });
