+++ apps/test-app/src/sidesheet/test-app-sidesheet/WorkorderSidesheet.tsx
@@ -1,8 +1,9 @@
 import { createWidget, useResizeContext } from '@equinor/workspace-sidesheet';
 import { Button } from '@equinor/eds-core-react';
 import { render as ScopeChangeRender } from './ScopeChangeSidesheet';
+import { sidesheets } from './sidesheets';
 
-type WorkOrderProps = {
+export type WorkOrderProps = {
 	woId: string;
 };
 
@@ -14,7 +15,7 @@ export const { Component: WorkOrderSidesheet, render } = createWidget<WorkOrderP
 			<button onClick={() => frame.unmount()}>unmount</button>
 			<div>Am Workorder sidesheet</div>
 			<div>My props are {JSON.stringify(props)}</div>
-			<Button onClick={() => frame.replace((el, frame) => ScopeChangeRender({ el: el, props: {}, frame }))}>
+			<Button onClick={() => frame.replace(sidesheets.scopechange({ id: 'some guid' }))}>
 				Replace with Scope change
 			</Button>
 		</div>
