+++ packages/sidesheet/src/lib/index.ts
@@ -1,2 +1,3 @@
 export * from './hooks';
 export * from './utils';
+export * from './types';
