+++ packages/sidesheet/src/lib/types/types.ts
@@ -1,0 +1,10 @@
+export type Cleanup = () => void;
+
+export type ReplaceFunction = (
+	newComp: (el: HTMLDivElement, replace: Frame) => Promise<Cleanup> | Cleanup
+) => Promise<Cleanup>;
+
+export type Frame = {
+	replace: ReplaceFunction;
+	unmount: () => void;
+};
